+++ packages/brex/src/types/vendor.ts
@@ -3,78 +3,78 @@ import * as S from '@effect/schema/Schema'
 /**
  * Payment Account Type
  */
-export const PaymentAccountType = S.literal(
+export const PaymentAccountType = S.Literal(
   'ACH',
   'CHEQUE',
   'DOMESTIC_WIRE',
   'INTERNATIONAL_WIRE'
 )
-export type PaymentAccountType = S.Schema.To<typeof PaymentAccountType>
+export type PaymentAccountType = S.Schema.Type<typeof PaymentAccountType>
 
 /**
  * Address
  */
 export const Address = S.partial(
-  S.struct({
-    line1: S.nullable(S.string),
-    line2: S.nullable(S.string),
-    city: S.nullable(S.string),
-    state: S.nullable(S.string),
-    country: S.nullable(S.string),
-    postal_code: S.nullable(S.string),
-    phone_number: S.nullable(S.string),
+  S.Struct({
+    line1: S.NullOr(S.String),
+    line2: S.NullOr(S.String),
+    city: S.NullOr(S.String),
+    state: S.NullOr(S.String),
+    country: S.NullOr(S.String),
+    postal_code: S.NullOr(S.String),
+    phone_number: S.NullOr(S.String),
   })
 )
-export type Address = S.Schema.To<typeof Address>
+export type Address = S.Schema.Type<typeof Address>
 
 /**
  * Payment Account Details
  */
-export const PaymentAccountDetails = S.struct({
+export const PaymentAccountDetails = S.Struct({
   type: PaymentAccountType,
-  payment_instrument_id: S.string,
-  routing_number: S.string,
-  account_number: S.string,
+  payment_instrument_id: S.String,
+  routing_number: S.String,
+  account_number: S.String,
   address: Address,
 })
-export type PaymentAccountDetails = S.Schema.To<typeof PaymentAccountDetails>
+export type PaymentAccountDetails = S.Schema.Type<typeof PaymentAccountDetails>
 
 /**
  * Vendor
  */
-const VendorRequired = S.struct({
-  id: S.string,
+const VendorRequired = S.Struct({
+  id: S.String,
 })
 const VendorOptional = S.partial(
-  S.struct({
-    company_name: S.string,
-    email: S.nullable(S.string),
-    phone: S.nullable(S.string),
-    payment_accounts: S.nullable(S.array(PaymentAccountDetails)),
+  S.Struct({
+    company_name: S.String,
+    email: S.NullOr(S.String),
+    phone: S.NullOr(S.String),
+    payment_accounts: S.NullOr(S.Array(PaymentAccountDetails)),
   })
 )
 export const Vendor = S.extend(VendorRequired, VendorOptional)
-export type Vendor = S.Schema.To<typeof Vendor>
+export type Vendor = S.Schema.Type<typeof Vendor>
 
 /**
  * Get Vendor Response
  */
 export const GetVendorResponse = Vendor
-export type GetVendorResponse = S.Schema.To<typeof GetVendorResponse>
+export type GetVendorResponse = S.Schema.Type<typeof GetVendorResponse>
 
 /**
  * List Vendors Response
  */
 const ListVendorsResponseOptional = S.partial(
-  S.struct({
-    next_cursor: S.string,
+  S.Struct({
+    next_cursor: S.String,
   })
 )
-const ListVendorsResponseRequired = S.struct({
-  items: S.array(Vendor),
+const ListVendorsResponseRequired = S.Struct({
+  items: S.Array(Vendor),
 })
 export const ListVendorsResponse = S.extend(
   ListVendorsResponseRequired,
   ListVendorsResponseOptional
 )
-export type ListVendorsResponse = S.Schema.To<typeof ListVendorsResponse>
+export type ListVendorsResponse = S.Schema.Type<typeof ListVendorsResponse>